+++ package.json
@@ -2,9 +2,10 @@
   "name": "server",
   "version": "1.0.0",
   "description": "yamovie server",
-  "main": "index.js",
+  "main": "server.js",
   "scripts": {
-    "start": "node index.js",
+    "start": "node server.js",
+    "dev": "nodemon server.js",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "repository": {
