+++ services/tmdb/index.js
@@ -1,13 +1,12 @@
-const request = require('request');
 const requestPromise = require('request-promise-native');
-const db = require('../data/db');
-const options = require('./tmdb-options');
-const parser = require('./tmdb-parser');
+const request = require('request');
+const db = require('../../data/db');
+const options = require('./options');
+const parser = require('./parser');
 
 let tmdb = {};
 
 tmdb.seedMovies = () => {
-  console.log('Seeding movies...');
   request(options.init, (err, res, body) => {
     if (err) throw new Error(err);
 
@@ -23,7 +22,6 @@ tmdb.seedMovies = () => {
 };
 
 tmdb.getMovieDetails = async id => {
-  console.log('getting movies from tmdb');
   const options = {
     method: 'GET',
     url: `https://api.themoviedb.org/3/movie/${id}`,
@@ -42,34 +40,9 @@ tmdb.getMovieDetails = async id => {
       .assign({ details })
       .write();
   });
-
-  // request(options, (err, res, body) => {
-  //   if (err) throw new Error(err);
-
-  //   const details = parser.details(body);
-
-  //   const movie = db
-  //     .get('movies')
-  //     .find({ id })
-  //     .assign({ details })
-  //     .write();
-  // });
 };
 
-// tmdb.getMovieDetails = async id => {
-//   const options = {
-//     method: 'GET',
-//     url: `https://api.themoviedb.org/3/movie/${id}`,
-//     qs: {
-//       api_key: process.env.TMDB_KEY
-//     },
-//     json: true,
-//     body: '{}'
-//   };
-// };
-
 tmdb.seedGenres = () => {
-  console.log('Seeding genres...');
   request(options.genres, (err, res, body) => {
     if (err) throw new Error(err);
